+++ spec/spec_helper.rb
@@ -11,14 +11,55 @@ def file(fname)
 	File.join(__dir__, "files", fname)
 end
 
-def xml(name)
-	file = "#{__dir__}/xml/#{name}.xml"
-	if File.exists?(file)
-		prefix = %|<?xml version="1.0" encoding="utf-8" standalone="no"?><omnifocus xmlns="http://www.omnigroup.com/namespace/OmniFocus/v2">|
-		suffix = "</omnifocus>"
-		data = prefix + File.read(file) + suffix
-		return Nokogiri::XML(data).at_xpath("/xmlns:omnifocus").elements.first
+# Super-powered XML method can produce XML from a file, a string, or a block
+def xml(file:nil, string:nil, &block)
+	xml_str = if file
+		path = "#{__dir__}/xml/#{file}.xml"
+		if File.exists?(path)
+			File.read(path)
+		else
+			raise ArgumentError, "XML file \"#{name}\" not recognised."
+		end
+	elsif string
+		string
+	elsif block
+		XMLMaker.make(&block)
 	else
-		raise ArgumentError, "XML file \"#{name}\" not recognised."
+		raise ArgumentError, "No argument supplied to xml()."
+	end
+
+	prefix = %|<?xml version="1.0" encoding="utf-8" standalone="no"?><omnifocus xmlns="http://www.omnigroup.com/namespace/OmniFocus/v2">|
+	suffix = "</omnifocus>"
+	xml_str = prefix + xml_str + suffix
+	return Nokogiri::XML(xml_str).at_xpath("/xmlns:omnifocus").elements.first
+end
+
+#---------------------------------------
+# XML maker
+module XMLMaker
+	def self.make(&blck)
+		instance_exec(&blck)
+	end
+	  
+	def self.tag(name, opts={}, &blck)
+		opts = opts.map{ |k,v| %| #{k}="#{v.gsub('"','\"')}"| }.join("")
+		if blck
+			"<#{name}#{opts}>" + instance_exec(&blck) + "</#{name}>"
+		else
+			"<#{name}#{opts} />"
+		end
 	end
 end
+
+# Will only puts if $debug is set
+def verbose(*str)
+	puts(*str) if $verbose
+end
+
+def verbosely &blck
+	$verbose = true
+	blck[]
+	$verbose = false
+end
+
+$verbose = false